Move appropriate properties from BTPayPalRequest to BTPayPalCheckoutRequest
