+++ Sources/BraintreePayPal/BTPayPalCheckoutRequest_Internal.h
@@ -1,0 +1,15 @@
+#if __has_include(<Braintree/BraintreePayPal.h>)
+#import <Braintree/BTPayPalCheckoutRequest.h>
+#else
+#import <BraintreePayPal/BTPayPalCheckoutRequest.h>
+#endif
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface BTPayPalCheckoutRequest ()
+
+@property (nonatomic, copy, readonly) NSString *intentAsString;
+
+@end
+
+NS_ASSUME_NONNULL_END
