+++ UnitTests/BraintreePayPalTests/BTPayPalRequest_Tests.swift
@@ -17,151 +17,31 @@ class BTPayPalRequest_Tests: XCTestCase {
         configuration = BTConfiguration(json: json)
     }
 
-    // MARK: - intentAsString
-
-    func testIntentAsString_whenIntentIsNotSpecified_returnsAuthorize() {
-        let request = BTPayPalRequest(amount: "1")
-        XCTAssertEqual(request.intentAsString, "authorize")
-    }
-
-    func testIntentAsString_whenIntentIsAuthorize_returnsAuthorize() {
-        let request = BTPayPalRequest(amount: "1")
-        request.intent = .authorize
-        XCTAssertEqual(request.intentAsString, "authorize")
-    }
-
-    func testIntentAsString_whenIntentIsSale_returnsSale() {
-        let request = BTPayPalRequest(amount: "1")
-        request.intent = .sale
-        XCTAssertEqual(request.intentAsString, "sale")
-    }
-
-    func testIntentAsString_whenIntentIsOrder_returnsOrder() {
-        let request = BTPayPalRequest(amount: "1")
-        request.intent = .order
-        XCTAssertEqual(request.intentAsString, "order")
-    }
-
     // MARK: - landingPageTypeAsString
 
     func testLandingPageTypeAsString_whenLandingPageTypeIsNotSpecified_returnNil() {
-        let request = BTPayPalRequest(amount: "1")
+        let request = BTPayPalRequest()
         XCTAssertNil(request.landingPageTypeAsString)
     }
 
     func testLandingPageTypeAsString_whenLandingPageTypeIsBilling_returnsBilling() {
-        let request = BTPayPalRequest(amount: "1")
+        let request = BTPayPalRequest()
         request.landingPageType = .billing
         XCTAssertEqual(request.landingPageTypeAsString, "billing")
     }
 
     func testLandingPageTypeAsString_whenLandingPageTypeIsLogin_returnsLogin() {
-        let request = BTPayPalRequest(amount: "1")
+        let request = BTPayPalRequest()
         request.landingPageType = .login
         XCTAssertEqual(request.landingPageTypeAsString, "login")
     }
 
-    // MARK: - parametersWithConfiguration (checkout flow)
-
-    func testParametersWithConfiguration_whenIsNotBillingAgreement_returnsAllParams() {
-        let request = BTPayPalRequest(amount: "1")
-        request.intent = .sale
-        request.offerCredit = true
-        request.offerPayLater = true
-        request.isShippingAddressRequired = true
-        request.displayName = "Display Name"
-        request.landingPageType = .login
-        request.localeCode = "locale-code"
-        request.merchantAccountId = "merchant-account-id"
-        request.currencyCode = "currency-code"
-
-        let shippingAddress = BTPostalAddress()
-        shippingAddress.streetAddress = "123 Main"
-        shippingAddress.extendedAddress = "Unit 1"
-        shippingAddress.locality = "Chicago"
-        shippingAddress.region = "IL"
-        shippingAddress.postalCode = "11111"
-        shippingAddress.countryCodeAlpha2 = "US"
-        shippingAddress.recipientName = "Recipient"
-        request.shippingAddressOverride = shippingAddress
-        request.isShippingAddressEditable = true
-
-        request.lineItems = [BTPayPalLineItem(quantity: "1", unitAmount: "1", name: "item", kind: .credit)]
-
-        let parameters = request.parameters(with: configuration, isBillingAgreement: false)
-        guard let experienceProfile = parameters["experience_profile"] as? [String : Any] else { XCTFail(); return }
-
-        XCTAssertEqual(parameters["intent"] as? String, "sale")
-        XCTAssertEqual(parameters["amount"] as? String, "1")
-        XCTAssertEqual(parameters["offer_paypal_credit"] as? Bool, true)
-        XCTAssertEqual(parameters["offer_pay_later"] as? Bool, true)
-        XCTAssertEqual(experienceProfile["no_shipping"] as? Bool, false)
-        XCTAssertEqual(experienceProfile["brand_name"] as? String, "Display Name")
-        XCTAssertEqual(experienceProfile["landing_page_type"] as? String, "login")
-        XCTAssertEqual(experienceProfile["locale_code"] as? String, "locale-code")
-        XCTAssertEqual(parameters["merchant_account_id"] as? String, "merchant-account-id")
-        XCTAssertEqual(parameters["currency_iso_code"] as? String, "currency-code")
-        XCTAssertEqual(experienceProfile["address_override"] as? Bool, false)
-        XCTAssertEqual(parameters["line1"] as? String, "123 Main")
-        XCTAssertEqual(parameters["line2"] as? String, "Unit 1")
-        XCTAssertEqual(parameters["city"] as? String, "Chicago")
-        XCTAssertEqual(parameters["state"] as? String, "IL")
-        XCTAssertEqual(parameters["postal_code"] as? String, "11111")
-        XCTAssertEqual(parameters["country_code"] as? String, "US")
-        XCTAssertEqual(parameters["recipient_name"] as? String, "Recipient")
-        XCTAssertEqual(parameters["line_items"] as? [[String : String]], [["quantity" : "1",
-                                                                                   "unit_amount": "1",
-                                                                                   "name": "item",
-                                                                                   "kind": "credit"]])
-
-        XCTAssertEqual(parameters["return_url"] as? String, "sdk.ios.braintree://onetouch/v1/success")
-        XCTAssertEqual(parameters["cancel_url"] as? String, "sdk.ios.braintree://onetouch/v1/cancel")
-    }
-
-    func testParametersWithConfiguration_whenIsNotBillingAgreement_andCurrencyCodeNotSet_usesConfigCurrencyCode() {
-        let json = BTJSON(value: [
-            "paypalEnabled": true,
-            "paypal": [
-                "currencyIsoCode": "currency-code"
-            ]
-        ])
-
-        configuration = BTConfiguration(json: json)
-
-        let request = BTPayPalRequest(amount: "1")
-        let parameters = request.parameters(with: configuration, isBillingAgreement: false)
-
-        XCTAssertEqual(parameters["currency_iso_code"] as? String, "currency-code")
-    }
-
-    func testParametersWithConfiguration_whenIsNotBillingAgreement_andShippingAddressIsRequiredNotSet_returnsNoShippingTrue() {
-        let request = BTPayPalRequest(amount: "1")
-        request.currencyCode = "GBP"
-        // no_shipping = true should be the default.
-
-        let parameters = request.parameters(with: configuration, isBillingAgreement: false)
-        guard let experienceProfile = parameters["experience_profile"] as? [String : Any] else { XCTFail(); return }
-
-        XCTAssertEqual(experienceProfile["no_shipping"] as? Bool, true)
-    }
-
-    func testParametersWithConfiguration_whenIsNotBillingAgreement_andShippingAddressIsRequiredIsTrue_returnsNoShippingFalse() {
-        let request = BTPayPalRequest(amount: "1")
-        request.currencyCode = "GBP"
-        request.isShippingAddressRequired = true
-
-        let parameters = request.parameters(with: configuration, isBillingAgreement: false)
-        guard let experienceProfile = parameters["experience_profile"] as? [String:Any] else { XCTFail(); return }
-        XCTAssertEqual(experienceProfile["no_shipping"] as? Bool, false)
-    }
-
-    // MARK: - parametersWithConfiguration (vault flow)
+    // MARK: - parametersWithConfiguration
 
     func testParametersWithConfiguration_whenIsBillingAgreement_returnsAllParams() {
         let request = BTPayPalRequest()
         request.billingAgreementDescription = "description"
         request.offerCredit = true
-        request.offerPayLater = true
         request.isShippingAddressRequired = true
         request.displayName = "Display Name"
         request.landingPageType = .login
@@ -187,7 +67,6 @@ class BTPayPalRequest_Tests: XCTestCase {
 
         XCTAssertEqual(parameters["description"] as? String, "description")
         XCTAssertEqual(parameters["offer_paypal_credit"] as? Bool, true)
-        XCTAssertEqual(parameters["offer_pay_later"] as? Bool, true)
         XCTAssertEqual(experienceProfile["no_shipping"] as? Bool, false)
         XCTAssertEqual(experienceProfile["brand_name"] as? String, "Display Name")
         XCTAssertEqual(experienceProfile["landing_page_type"] as? String, "login")
@@ -210,39 +89,22 @@ class BTPayPalRequest_Tests: XCTestCase {
         XCTAssertEqual(parameters["cancel_url"] as? String, "sdk.ios.braintree://onetouch/v1/cancel")
     }
 
-    func testParametersWithConfiguration_whenIsBillingAgreement_doesNotReturnIntent() {
+    func testParametersWithConfiguration_whenShippingAddressIsRequiredNotSet_returnsNoShippingTrue() {
         let request = BTPayPalRequest()
-        request.intent = .authorize
+        // no_shipping = true should be the default.
 
-        let parameters = request.parameters(with: configuration, isBillingAgreement: true)
+        let parameters = request.parameters(with: configuration, isBillingAgreement: false)
+        guard let experienceProfile = parameters["experience_profile"] as? [String : Any] else { XCTFail(); return }
 
-        XCTAssertNil(parameters["intent"])
+        XCTAssertEqual(experienceProfile["no_shipping"] as? Bool, true)
     }
 
-    func testParametersWithConfiguration_whenIsBillingAgreement_andConfigurationHasCurrency_doesNotReturnCurrency() {
-        let json = BTJSON(value: [
-            "paypalEnabled": true,
-            "paypal": [
-                "environment": "offline",
-                "currencyIsoCode": "GBP",
-            ] ])
-
-        configuration = BTConfiguration(json: json)
-
+    func testParametersWithConfiguration_whenShippingAddressIsRequiredIsTrue_returnsNoShippingFalse() {
         let request = BTPayPalRequest()
+        request.isShippingAddressRequired = true
 
-        let parameters = request.parameters(with: configuration, isBillingAgreement: true)
-
-        XCTAssertNil(parameters["currency_iso_code"])
-    }
-
-    func testParametersWithConfiguration_whenIsBillingAgreement_andRequestHasCurrency_doesNotReturnCurrency() {
-        let request = BTPayPalRequest()
-        request.currencyCode = "GBP"
-
-        let parameters = request.parameters(with: configuration, isBillingAgreement: true)
-
-        XCTAssertNil(parameters["currency_iso_code"])
-        XCTAssertNil(parameters["intent"])
+        let parameters = request.parameters(with: configuration, isBillingAgreement: false)
+        guard let experienceProfile = parameters["experience_profile"] as? [String:Any] else { XCTFail(); return }
+        XCTAssertEqual(experienceProfile["no_shipping"] as? Bool, false)
     }
 }
